made views into 2 files
