+++ tutor/views/__init__.py
@@ -1,0 +1,2 @@
+from tutor.views.main import *
+from tutor.views.gapi import *
