+++ tutor/views/gapi.py
@@ -1,46 +1,21 @@
+"""
+gapi.py
+This is the file for the views that use the google api
+"""
 import httplib2
 import re
 import os
-from django.shortcuts import render, redirect
-from django.http import HttpResponse, Http404
-from django.urls import reverse
-from django.contrib.auth.models import User
-from django.contrib.auth.decorators import login_required
 from base64 import urlsafe_b64encode
 from apiclient import discovery
 from oauth2client import client
 
+from django.shortcuts import render, redirect
+from django.http import HttpResponse
+from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
+from django.urls import reverse
+
 import tutor.models as models
-
-
-def index(request):
-    return render(
-        request,
-        'tutor/index.html',
-        {"Courses": list(models.Course.objects.all())})
-
-
-def tutors(request, course_id):
-    """
-    View for the tutors page
-    It gets the course based on the course id put in the address bar.
-    Then, if the course exists, it renders the tutors.html page and passes
-    the list of tutors on that course to the page.
-    """
-    try:
-        course = models.Course.objects.get(pk=course_id)
-    except models.Course.DoesNotExist:
-        raise Http404("Course does not exist")
-    # student_set has the info of all of the users who are marked as tutors
-    # for this course
-    return render(
-        request,
-        'tutor/tutors.html',
-        {"tutors": course.student_set.all()})
-
-
-def startstop(request):
-    return render(request, 'tutor/startstop.html')
 
 
 @login_required(login_url='/admin/login/')
